--- common_msgs/msg/TimePose.msg
@@ -1,3 +1,0 @@
-time timestamp
-geometry_msgs/Pose2D pose
-
