+++ common_msgs/srv/AddTwoNum.srv
@@ -1,4 +1,4 @@
 int32 a
 int32 b
 ---
-int32 mul
+int32 sum
